Prevent mousedown -> drag -> mouseup from creating clicks
Using the mouse, dragging an element causes a click, as the down and up events happen
at the same dom element (which moves under the mouse). This feels very bad and causes
accidental clicks. The fix is to add a click listener to the capture phase and prevent
clicks where:
- the interaction started within react-touch-carousel
- the mouse moved more than 5px ( or value specified by user)